rename alloc feature mod
