+++ src/features/mod.rs
@@ -2,7 +2,7 @@
 mod serde;
 
 #[cfg(feature = "nightly-allocator-api")]
-pub mod alloc;
+pub mod nightly_allocator_api;
 
 #[cfg(feature = "allocator-api2-02")]
 pub mod allocator_api2_02;
